requirements.txt
requests>=2.31.0
anthropic>=0.7.0
python-dotenv>=1.0.0
rich>=13.0.0
jupyter>=1.0.0
notebook>=6.0.0